Переименование компонента MeetsList на MeetListPage
